python_backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
transformers==4.41.2
torch==2.3.1
Pillow==10.3.0
python-multipart==0.0.9
